requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
faster-whisper==0.10.0
google-cloud-translate==3.12.1
transformers==4.37.2
torch==2.2.0
apify-client==1.6.3
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.6.1
aiohttp==3.9.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
websockets==12.0
aiosmtplib==3.0.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
huggingface-hub==0.20.3 